implement logExtcutionTime & update enitity & delete method

+++ src/main/java/buloshnaya/orders/kafka/dto/OrderItemDto.java
@@ -1,8 +1,10 @@
 package buloshnaya.orders.kafka.dto;
 
+import java.math.BigDecimal;
+
 public record OrderItemDto(
     String productId,
     String productName,
-    Integer price,
+    BigDecimal price,
     Integer quantity
 ) {}
